Restructering halt/reboot files
The function halt() and reboot() used to be stored
in two different c files and two different header
files which were named differently than the c files.
This commit merges those two functions into the
same file and provides an unique header file (shutdown.h)
for accessing them.
