+++ kernel/arch/sparc64/src/smp/sun4v/smp.c
@@ -42,7 +42,6 @@
 #include <arch.h>
 #include <config.h>
 #include <macros.h>
-#include <halt.h>
 #include <stdbool.h>
 #include <stddef.h>
 #include <stdint.h>
