+++ kernel/generic/src/console/cmd.c
@@ -51,7 +51,6 @@
 #include <adt/list.h>
 #include <arch.h>
 #include <config.h>
-#include <halt.h>
 #include <str.h>
 #include <macros.h>
 #include <cpu.h>
